requirements.txt
langchain
langchain-openai
python-dotenv
pydantic
langchain_core
langchain-groq